requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
sentence-transformers==3.1.1
pinecone-client==5.0.1
pdfminer.six==20231228
python-docx==1.1.2
requests==2.32.3
numpy==1.26.4
google-generativeai==0.8.3

